+++ CHANGELOG.md
@@ -1,12 +1,15 @@
-#### 0.3.8 (2018-12-03)
+#### 0.3.8 (2018-12-04)
 
 ##### Documentation Changes
 
+*  Build docs for orders ([b7c47397](https://github.com/big-neon/bn-api-node/commit/b7c473970bd5c6dc2b42375be3e176dc6c39c46b))
 *  Build docs for comps ([c0ca726e](https://github.com/big-neon/bn-api-node/commit/c0ca726e0eadf040b5f2594964d2d23ddafa2a87))
 
 ##### New Features
 
 * **endpoints:**
+  *  Added `orders.tickets.index` and `orders.update` ([6018d53f](https://github.com/big-neon/bn-api-node/commit/6018d53f9f9066420c57353fb813a6fc827c7dcf))
+  *  Added /events/checkins endpoint ([#238](https://github.com/big-neon/bn-api-node/pull/238)) ([58017291](https://github.com/big-neon/bn-api-node/commit/5801729119523fab9404045ade99b2f30edcdbe4))
   *  comps.[update, del, read] endpoints ([1c3e8a5d](https://github.com/big-neon/bn-api-node/commit/1c3e8a5d35b9888a25960073f2ab27e16d0b9d79))
   *  Added /events/checkins endpoint ([#238](https://github.com/big-neon/bn-api-node/pull/238)) ([58017291](https://github.com/big-neon/bn-api-node/commit/5801729119523fab9404045ade99b2f30edcdbe4))
   *  Added organizations.fans.[read, history, index] ([0c4f35f7](https://github.com/big-neon/bn-api-node/commit/0c4f35f7b0307127d4ac3f3681495f6198d65173))